pyproject.toml
[project]
name = "blog-rss-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.43",
    "feedgen>=1.0.0",
    "playwright>=1.55.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "typer>=0.19.2",
]
